+++ R/compliment2-package.R
@@ -1,0 +1,10 @@
+#' @import rlang
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+
+## mockable bindings: start
+## mockable bindings: end
+NULL