+++ NEWS.md
@@ -1,5 +1,6 @@
 # compliment2 0.1.0
 
 * New `compliment_r()` generates random compliments about coding in R.
+* New `farewell()` generates random goodbyes.
 
 * Initial CRAN submission.